Refactoring: 1. User logic moved to separate module, 2. Numerous imports wrapper up into shorter statements.
