+++ tasker/logging.py
@@ -1,0 +1,5 @@
+def user_logout(request):
+    if request.method == 'POST':
+        if request.user.is_authenticated:
+            logout(request)
+    return redirect('/accounts/logged_out.html')
